can now arrange order of tasks, but doesn't persist

+++ ToDoList - Assignment 3/app/src/main/java/com/snerdette/needylist/DashboardActivity.kt
@@ -16,6 +16,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.snerdette.needylist.DTO.ToDo
+import com.snerdette.needylist.DTO.ToDoItem
 import kotlinx.android.synthetic.main.activity_dashboard.*
 
 class DashboardActivity : AppCompatActivity() {
@@ -108,8 +109,16 @@ class DashboardActivity : AppCompatActivity() {
                             activity.updateToDo(list[p1])
                         }
                         R.id.menu_delete->{
-                            activity.dbHandler.deleteToDo(list[p1].id)
-                            activity.refreshList()
+                            val dialog = AlertDialog.Builder(activity)
+                            dialog.setTitle("Are you sure")
+                            dialog.setMessage("Do you want to delete this task?")
+                            dialog.setPositiveButton("Continue") { _: DialogInterface, _: Int ->
+                                activity.dbHandler.deleteToDo(list[p1].id)
+                                activity.refreshList()
+                            }
+                            dialog.setNegativeButton("Cancel") { _: DialogInterface, _: Int ->
+                            }
+                            dialog.show()
                         }
                         R.id.menu_mark_as_completed->{
                             activity.dbHandler.updateToDoItemCompletedStatus(list[p1].id,  true)
